+++ CHANGELOG.md
@@ -39,6 +39,9 @@ Line wrap the file at 100 chars.                                              Th
 ### Fixed
 - Fix so changing the OpenVPN mssfix setting triggers setting up a new tunnel with the new setting.
 
+#### macOS
+- Correctly backup and restore search domains when manipulating DNS settings.
+
 
 ## [2018.4-beta3] - 2018-10-12
 ### Fixed